Support bookmarks in vterm.
This patch allows saving bookmarks in vterm mode. It saves the
current buffer name, as well as the current directory.

When asked to restore a bookmark, it will create a new vterm buffer
if needed, and navigate to the previous directory if the custom
variable vterm-bookmark-check-dir is set to non-nil.

This can be quite convenient for people who constantly have several
vterm buffers with different names/dirs for different needs.